+++ apps/router-demo/router-remote4-2004/CHANGELOG.md
@@ -1,5 +1,22 @@
 # remote4
 
+## 1.0.19
+
+### Patch Changes
+
+- Updated dependencies [f245bb3]
+- Updated dependencies [1d9bb77]
+  - @module-federation/enhanced@0.6.0
+  - @module-federation/bridge-react@0.6.0
+
+## 1.0.18
+
+### Patch Changes
+
+- Updated dependencies [b90fa7d]
+  - @module-federation/enhanced@0.5.2
+  - @module-federation/bridge-react@0.5.2
+
 ## 1.0.17
 
 ### Patch Changes